Fix Interrupt bug
The interrupt would panic in the same stack that caused the error.
This meant another package up the stack could potentially try to
recover the error with it's own panic/recover logic - which would
not work. Instead, the panic happens in a go routine, guarenteeing
there is no recover above it in the stack. A thread is used to
block any further execution.