Added XchgPair to encapsulate public and private exchange keys
A private key is almost never useful alonge, so it made more
sense to store both in a struct together. By embedding both of
them, it's easy to use the key as either.

Swapped the order on computing a shared key, it is now a method
on Priv rather than Pub. This makes it cleaner to use with a
keypair.